requirements.txt
click==8.0.1
Flask==2.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib